WIP cypress tests respond to PR comments
