+++ apps/coscrad-frontend/src/components/resources/utils/render-aggregate-id-cell.tsx
@@ -1,9 +1,10 @@
 import { HasId } from '@coscrad/api-interfaces';
+import { Button } from '@mui/material';
 import { Link } from 'react-router-dom';
 import { CellRenderer } from '../../../utils/generic-components/presenters/tables/generic-index-table-presenter/types';
 
 export const renderAggregateIdCell: CellRenderer<HasId> = ({ id }: HasId) => (
     <Link data-testid={id} to={id}>
-        VIEW
+        <Button variant="outlined">VIEW</Button>
     </Link>
 );
